requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
paho-mqtt==2.1.0
redis==5.2.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
python-dotenv==1.0.1
jinja2==3.1.5
apscheduler==3.10.4
